refactor: move schemas to separate files
