+++ src/probe/handler/dns.ts
@@ -1,10 +1,8 @@
-import Joi from 'joi';
+import { dnsSchema } from '../schema/probe-response-schema.js';
 import type { Probe } from '../types.js';
 
-const schema = Joi.array<string[]>().items(Joi.string()).required();
-
 export const handleDnsUpdate = (probe: Probe) => (list: string[]): void => {
-	const validation = schema.validate(list);
+	const validation = dnsSchema.validate(list);
 
 	if (validation.error) {
 		throw validation.error;
