+++ src/probe/handler/status.ts
@@ -1,10 +1,8 @@
-import Joi from 'joi';
+import { statusSchema } from '../schema/probe-response-schema.js';
 import type { Probe } from '../types.js';
 
-const schema = Joi.string<Probe['status']>().valid('initializing', 'ready', 'unbuffer-missing', 'ping-test-failed', 'sigterm').required();
-
 export const handleStatusUpdate = (probe: Probe) => (status: Probe['status']) => {
-	const validation = schema.validate(status);
+	const validation = statusSchema.validate(status);
 
 	if (validation.error) {
 		throw validation.error;
